+++ test/smokes/remark_lint/expectations.rb
@@ -8,6 +8,17 @@ s.add_test(
   issues: [
     {
       path: "readme.md",
+      location: { start_line: 7, start_column: 12 },
+      id: "missing-heading",
+      message: "Link to unknown heading: `foo`",
+      links: [],
+      object: { severity: "warn" },
+      git_blame_info: {
+        commit: :_, line_hash: "46f054a009c80f7f5c6cd602038c951d186c2276", original_line: 7, final_line: 7
+      }
+    },
+    {
+      path: "readme.md",
       location: { start_line: 3, start_column: 1 },
       id: "no-undefined-references",
       message: "Found reference to undefined definition",
@@ -215,7 +226,12 @@ s.add_test(
   analyzer: { name: "remark-lint", version: default_version }
 )
 
-s.add_test("with_remark_lint_package", type: "success", issues: [], analyzer: { name: "remark-lint", version: default_version })
+s.add_test(
+  "with_remark_lint_package",
+  type: "success",
+  issues: [],
+  analyzer: { name: "remark-lint", version: default_version }
+)
 
 s.add_test(
   "without_remark_lint_package",
@@ -230,6 +246,44 @@ s.add_test(
       object: { severity: "warn" },
       git_blame_info: {
         commit: :_, line_hash: "18c2ad20474bed96b6a042d5cbc1aec43ae7b65a", original_line: 1, final_line: 1
+      }
+    }
+  ],
+  analyzer: { name: "remark-lint", version: default_version }
+)
+
+s.add_test(
+  "use_git",
+  type: "success",
+  issues: [
+    {
+      path: "foo.md",
+      location: { start_line: 1, start_column: 1 },
+      id: "missing-file",
+      message: "Link to unknown file: `readme.md`",
+      links: [],
+      object: { severity: "warn" },
+      git_blame_info: {
+        commit: :_, line_hash: "a6f65463dbaa25c99a03f499f0e675d62db5e1e7", original_line: 1, final_line: 1
+      }
+    }
+  ],
+  analyzer: { name: "remark-lint", version: default_version }
+)
+
+s.add_test(
+  "use_git_root_dir",
+  type: "success",
+  issues: [
+    {
+      path: "src/foo.md",
+      location: { start_line: 1, start_column: 1 },
+      id: "missing-file",
+      message: "Link to unknown file: `readme.md`",
+      links: [],
+      object: { severity: "warn" },
+      git_blame_info: {
+        commit: :_, line_hash: "a6f65463dbaa25c99a03f499f0e675d62db5e1e7", original_line: 1, final_line: 1
       }
     }
   ],